requirements.txt
django == 2.0.4
django-ckeditor == 5.4.0
djangorestframework == 3.8.1
pytz == 2018.3